factor out oracle filters so they can be used in ethcall engine
