+++ core/evtforward/ethcall/call_test.go
@@ -7,6 +7,7 @@ import (
 
 	"code.vegaprotocol.io/vega/core/evtforward/ethcall"
 	"code.vegaprotocol.io/vega/core/types"
+	v1 "code.vegaprotocol.io/vega/protos/vega/data/v1"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -85,4 +86,53 @@ func TestContractCall2(t *testing.T) {
 	assert.Equal(t, args, res.Values)
 
 	assert.Equal(t, map[string]string{"static": "66"}, res.Normalised)
+}
+
+func TestContractFilters(t *testing.T) {
+	ctx := context.Background()
+	tc, err := NewToyChain()
+	require.NoError(t, err)
+
+	args := []any{
+		int64(42),
+		big.NewInt(42),
+		"hello",
+		true,
+		common.HexToAddress("0xb794f5ea0ba39494ce839613fffba74279579268"),
+	}
+
+	argsJson, err := ethcall.AnyArgsToJson(args)
+	require.NoError(t, err)
+
+	spec := types.EthCallSpec{
+		ArgsJson:    argsJson,
+		Address:     tc.contractAddr.Hex(),
+		AbiJson:     tc.abiBytes,
+		Method:      "testy1",
+		Normalisers: map[string]string{"badger": `$[0]`, "static": "66"},
+		Filters: []*types.DataSourceSpecFilter{
+			{
+				Key: &types.DataSourceSpecPropertyKey{
+					Name: "badger",
+					Type: v1.PropertyKey_TYPE_INTEGER,
+				},
+				Conditions: []*types.DataSourceSpecCondition{
+					{
+						Operator: v1.Condition_OPERATOR_GREATER_THAN_OR_EQUAL,
+						Value:    "50",
+					},
+				},
+			},
+		},
+	}
+
+	call, err := ethcall.NewCall(spec)
+	require.NoError(t, err)
+
+	res, err := call.Call(ctx, tc.client, 1)
+	require.NoError(t, err)
+	assert.NotEmpty(t, res.Bytes)
+	assert.False(t, res.PassesFilters)
+	assert.Equal(t, []any{int64(42), big.NewInt(42), "hello", true, common.HexToAddress("0xb794f5ea0ba39494ce839613fffba74279579268")}, res.Values)
+	assert.Equal(t, map[string]string{"badger": "42", "static": "66"}, res.Normalised)
 }
